+++ src/utils/runtime/src/prof.rs
@@ -1,0 +1,29 @@
+fn spawn_prof_thread(profile_path: String) -> std::thread::JoinHandle<()> {
+    tracing::info!("writing prof data to directory {}", profile_path);
+    let profile_path = PathBuf::from(profile_path);
+
+    std::fs::create_dir_all(&profile_path).unwrap();
+
+    std::thread::spawn(move || {
+        let mut cnt = 0;
+        loop {
+            let guard = pprof::ProfilerGuardBuilder::default()
+                .blocklist(&["libc", "libgcc", "pthread", "vdso"])
+                .build()
+                .unwrap();
+            std::thread::sleep(Duration::from_secs(60));
+            match guard.report().build() {
+                Ok(report) => {
+                    let profile_svg = profile_path.join(format!("{}.svg", cnt));
+                    let file = std::fs::File::create(&profile_svg).unwrap();
+                    report.flamegraph(file).unwrap();
+                    tracing::info!("produced {:?}", profile_svg);
+                }
+                Err(err) => {
+                    tracing::warn!("failed to generate flamegraph: {}", err);
+                }
+            }
+            cnt += 1;
+        }
+    })
+}
